requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
cryptg==0.3.1
dateparser==1.1.1
debugpy==1.3.0
decorator==5.0.9
frozenlist==1.3.0
idna==3.3
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
multidict==6.0.2
numpy==1.23.0
oauthlib==3.2.0
pandas==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
Pygments==2.9.0
PyJWT==2.4.0
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.1.0
regex==2022.3.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
Telethon==1.24.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tweepy==4.10.0
twilio==7.9.3
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.3
yarl==1.7.2
